Add new design for 2018 Congress
This commit adds a new design for the 2018 Congress site. Since it looks
like previous year-to-year changes have been very minor, I've attempted
to establish a few conventions that will make it easy for upcoming years
to modify as little or as much as they need to.

* Use a HAML render convention to grab a year-specific partial if
  present, and fall back on the general one if not.
* Begin storing year-specific assets, stylesheets, etc in YYYY folders
* Add a breakpoint mixin to make it easy to use responsive styling...
  we want the site to look good on mobile browsers

This is my first time working with Ruby or Rails, so please let me know
if I've missed anything obvious or failed to follow any important
conventions.