Rewrite `button_to_get` helper
We've had a couple reports of the "Add Adult Attendee" button not doing
anything. It's using kind of an unnecessarily complicated means
involving JS to act as a button that links to a URL with a query
string--just using CSS to make an anchor tag look like a button is
simpler, and should eliminate any possible issue with the button.

I've also fixed the issue with the attendee box bumping directly into
the buttons beneath it.