Change select box to radio button for single plan categories
When using radio buttons the name attribute must be the same for all
plans in the plan category. This causes the params for the plans with
radio buttons to differ from those with check boxes. I added a method to
covert the plan params using radio buttons to use the expected checkbox
format.