Add a refund_policy field to years model
Our refund/cancellation policy was one of the first things a user asked
for this year when registration opened. Which makes sense! This is
something every year's Congress should be prompted to set, and should
always show on the "Make a Payment" page.

It also adds a new "policies" box of settings to the admin, which will
be a good place to add any always-relevant policy fields to help each
year's Congress kick things off.