Add `minor_agreement_received` checkbox for admin
Closes #79.

I started to create a migration to add this checkbox, which has been
requested by our registrar and a previous registrar, only to discover
that it already existed. It looks like there used to be an issue with
the way the project originally protected admin-only attributes in the
registration form.

This restores these fields, then adds some new logic to the registration
controller to ensure that only admins can update these fields.

This also adds a (very simple) report to the `rpt` namespace to show a
list of minor attendees for whom we've got an agreement and those for
whom we don't.

* Adds a test to verify that admins can update admin-only fields in reg
* Adds a test of the new report controller