Add table of contents option to content categories
Some content categories--I'm looking at you, FAQ!--really need a nice
table of contents at the top to let people quickly scan all the items
and click particular ones that interest them.

This creates a new column attached to content_categories that turns on a
TOC and outputs it as a list of links at the top of the page in the
view.