Remove check for user balance on index
It turns out that checking a large (even modestly large) number of users
for their current balance created immense slowdowns (N + 1 problem).
There are some possible optimization strategies we can try down the
road, but the quickest solution for the moment is to abandon the check
for the index and only check individual users. This isn't a huge
problem, as it was more convenience to let people check in users
directly from the index than necessity.