Temporarily add static information to "Who's Coming"
Since this year's Congress (2019) is using a separate system for
registration, the dynamic "Who's Coming?" page isn't accurate. There's
work purportedly underway in syncing them, but we needed to put up
information in the meantime. This adds a static table generated
separately from a spreadhseet and comments out the page's normal dynamic
functionality.