Modify registration to get values from Members API
Make adding an attendee easier by searching the AGA Members API for
their name or AGA ID and using the values there to pre-fill parts of
their Congress registration.

The new flow presents a starting screen after clicking "add attendee"
that displays only a search box for the API and a button for anyone
trying to register without being a member. (Many non-players and other
guests are attendees at a typical Congress.)

Once either a member is selected via the search or the not-a-member
button is clicked, the next page is the previous registration form, but
with as many values as I could find that were relevant filled in from
existing info.