Run db migration when deploying to Heroku
See:
https://mentalized.net/journal/2017/04/22/run-rails-migrations-on-heroku-deploy/

--
And just like that, your database is migrated before you release the
new version to the public. The benefits of this are plentiful:

1. We cannot forget to run the migrations
2. The database has been migrated when your application is released
3. Migrations can be run by anyone/anything that can push to the git remote
4. If a migration fails, the deployment is halted, your database changes
are rolled back (you did use transactions, right?), and your application
keeps running.

This is great for continous deployment and a stable deployment process
in general.
--