Fix enum value
Changes the 'in-person' value to 'in_person'.

The use of hyphens in enums is problematic. For example, a method named
`in-person?` will be defined, but that method can't be called easily because the
hyphen is treated as an operator.

AFAIK, it is OK to change enum values as long as you don't change their order.
In this case, the 'in_person' value has the same integer value, 0, as its
predecesor.