Change "cancelled" to "canceled"
Both spellings are correct, so there's no fundamental judgment here, but
"canceled" is much more commonly accepted in the U.S., and "cancelled"
is much  more commonly accepted in British English.

@see: https://www.grammarly.com/blog/canceled-vs-cancelled/

At this time, I'm not updating all the uses in code, just all the places
it's displayed for the actual user.