Restore XML file format and AGA info
This restores the logic necessary to generate the XML files for player
list imports to OpenGotha. I also restored the information from the AGA
TD List. We'll want to refactor that part of the site to use info from
the Membership API instead of manually scraping the TD list, but it can
stay as-is for now.