Refactor who is coming list to fix minor listing
I introduced a regression into the "Who's Coming" list when I updated
the SQL query to accomodate online Congresses, which don't have the same
need to only include people who have paid a certain amount toward their
total cost.

In working on the fix, I realized that it's also a problem to have a
list of total attendees that does not include minors when reporting a
count. We want an actual count, but we also want a public list that
removes those minors for privacy reasons.

This refactor takes away any query-based removal of minors, and instead
offers a class-based method for returning a minor-removed public list of
attendees.