Change activity notes field to "text" type without character limit
The only real reason to have a limit is to keep things that are in the
registration workflow sane for users, but we already cut this
description off in the registration form and offer a "more" button to
link to the full activity. That means the activity really can be as long
as it wants.

This changes the field from a string type (which requires a character
limit) to a text type, which does not.

@see:
https://stackoverflow.com/questions/8694273/changing-a-column-type-to-longer-strings-in-rails

Closes #259