Set minimum_coverage 81
Setting a minimum prevents backsliding. At the end of the test suite,
simplecov will exit with nonzero status, alerting the reviewer to the
problem.

I'd recommend 90 as a reasonable goal. We can ratchet up the minimum
as we go. I've had projects at work with coverage as high as 94, but
there are diminishing returns as you get closer to 100.