Fix Registration#attendee_number
The definition was changed in 5f1195b0e5b40db45e2d6e69a58570f84684bd41,
which broke two tests, e.g. registrations_controller_spec.rb:73

I believe the change was incorrect, as explained in my comment. So,
this commit reverts that change. This is a *partial* reversion of
5f1195b0e5b40db45e2d6e69a58570f84684bd41. The other parts are not
familiar to me (I don't know what "canceling" an attendee means)
so I left them as is.