Add option to export player list with only usernames
For online games during the e-Go Congress, many players do not wish to
have their full names matched with their usernames. This adds the option
to export a file for OpenGotha that has the player's first name as their
server username, and their last name listed as "X", since OpenGotha
requires a value.