Add checkbox to accept Covid-19 policy
In another feature that I hope will be utterly useless for future
Congresses, this adds a setting that allows an admin to add a link to a
covid_policy_url. If that link is present, it adds a required checkbox
to registration with a link to that policy that requires the user to
indicate that they understand and accept it.